+++ far2l/src/vmenu.hpp
@@ -265,7 +265,7 @@ private:
 	int ItemHiddenCount;
 	int ItemSubMenusCount;
 
-	BYTE Colors[VMENU_COLOR_COUNT];
+	uint64_t Colors[VMENU_COLOR_COUNT];
 
 	int MaxLineWidth;
 	bool bRightBtnPressed;
@@ -335,7 +335,7 @@ public:
 
 	void SetColors(struct FarListColors *ColorsIn = nullptr);
 	void GetColors(struct FarListColors *ColorsOut);
-	void SetOneColor(int Index, short Color);
+	void SetOneColor(int Index, uint64_t Color);
 
 	virtual int ProcessKey(FarKey Key);
 	virtual int ProcessMouse(MOUSE_EVENT_RECORD *MouseEvent);
